commandHauls: extract prompt to module
